Change widget id to dropdownReference / dropdownReferenceSet
