+++ Gruntfile.js
@@ -2,7 +2,7 @@
 const webpack = require("webpack");
 const webpackConfig = require("./webpack.config");
 const merge = require("webpack-merge");
-const widgetNames = [ "DropdownTypeahead", "DropdownTypeaheadReferenceSet" ];
+const widgetNames = [ "DropdownReference", "DropdownReferenceSet" ];
 
 const webpackConfigRelease = webpackConfig.map(config => merge(config, {
     devtool: false,
