+++ src/package.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <package xmlns="http://www.mendix.com/package/1.0/">
-    <clientModule name="DropdownTypeahead" version="1.0.0" xmlns="http://www.mendix.com/clientModule/1.0/">
+    <clientModule name="Dropdown" version="1.0.0" xmlns="http://www.mendix.com/clientModule/1.0/">
         <widgetFiles>
-            <widgetFile path="DropdownTypeahead/DropdownTypeahead.xml"/>
-            <widgetFile path="DropdownTypeaheadReferenceSet/DropdownTypeaheadReferenceSet.xml"/>
+            <widgetFile path="DropdownReference/DropdownReference.xml"/>
+            <widgetFile path="DropdownReferenceSet/DropdownReferenceSet.xml"/>
         </widgetFiles>
         <files>
-            <file path="com/mendix/widget/custom/DropdownTypeahead"/>
-            <file path="com/mendix/widget/custom/DropdownTypeaheadReferenceSet"/>
+            <file path="com/mendix/widget/custom/dropdown/DropdownReference"/>
+            <file path="com/mendix/widget/custom/dropdown/DropdownReferenceSet"/>
         </files>
     </clientModule>
 </package>
