+++ tests/e2e/wdio.conf.js
@@ -14,8 +14,8 @@ exports.config = {
     coloredLogs: true,
     bail: 0,
     screenshotPath: "dist/wdio/",
-    baseUrl: debug ? "http://localhost:8081/" : "https://dropdowntypeahead.mxapps.io/",
-    waitforTimeout: 3600,
+    baseUrl: debug ? "http://localhost:8080/" : "https://dropdowntypeahead.mxapps.io/",
+    waitforTimeout: 36000,
     connectionRetryTimeout: 200000,
     connectionRetryCount: 2,
     services: [ "selenium-standalone" ],
